REPORT
The Report consist of the Abstract, Problem in maintaining and managing the library i.e. using Manual records And Solution to the problem faced in maintaining manual data by Developing the Electronic Management System for managing the data and tasks electronically i.e. on the Computer.
It also consist of the different window pages of the Electronic Management System and the respective code of the respective pages